Moved and updated GET, PUT and DELETE request functions
